+++ instrumentation/github.com/gin-gonic/gin/otelgin/gin.go
@@ -31,6 +31,7 @@ const (
 // server handling the request.
 func Middleware(service string, opts ...Option) gin.HandlerFunc {
 	cfg := config{}
+
 	for _, opt := range opts {
 		opt.apply(&cfg)
 	}
@@ -95,6 +96,8 @@ func Middleware(service string, opts ...Option) gin.HandlerFunc {
 			oteltrace.WithSpanKind(oteltrace.SpanKindServer),
 		}
 
+		opts = append(opts, cfg.SpanStartOptions...)
+
 		spanName := cfg.SpanNameFormatter(c)
 		if spanName == "" {
 			spanName = fmt.Sprintf("HTTP %s route not found", c.Request.Method)
@@ -122,6 +125,9 @@ func Middleware(service string, opts ...Option) gin.HandlerFunc {
 
 		// Record the server-side attributes.
 		var additionalAttributes []attribute.KeyValue
+		if c.FullPath() != "" {
+			additionalAttributes = append(additionalAttributes, sc.Route(c.FullPath()))
+		}
 		if cfg.MetricAttributeFn != nil {
 			additionalAttributes = append(additionalAttributes, cfg.MetricAttributeFn(c.Request)...)
 		}