+++ otelconf/v0.3.0/log_test.go
@@ -5,13 +5,21 @@ package otelconf
 
 import (
 	"context"
+	"crypto/tls"
+	"crypto/x509"
 	"errors"
+	"net"
+	"os"
 	"path/filepath"
 	"reflect"
+	"strings"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials"
 
 	"go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploggrpc"
 	"go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploghttp"
@@ -19,7 +27,9 @@ import (
 	"go.opentelemetry.io/otel/log"
 	"go.opentelemetry.io/otel/log/noop"
 	sdklog "go.opentelemetry.io/otel/sdk/log"
+	sdklogtest "go.opentelemetry.io/otel/sdk/log/logtest"
 	"go.opentelemetry.io/otel/sdk/resource"
+	collogpb "go.opentelemetry.io/proto/otlp/collector/logs/v1"
 )
 
 func TestLoggerProvider(t *testing.T) {
@@ -693,4 +703,155 @@ func TestLogProcessor(t *testing.T) {
 			}
 		})
 	}
+}
+
+func Test_otlpGRPCLogExporter(t *testing.T) {
+	type args struct {
+		ctx        context.Context
+		otlpConfig *OTLP
+	}
+	tests := []struct {
+		name           string
+		args           args
+		grpcServerOpts func() ([]grpc.ServerOption, error)
+	}{
+		{
+			name: "no TLS config",
+			args: args{
+				ctx: context.Background(),
+				otlpConfig: &OTLP{
+					Protocol:    ptr("grpc"),
+					Compression: ptr("gzip"),
+					Timeout:     ptr(5000),
+					Insecure:    ptr(true),
+					Headers: []NameStringValuePair{
+						{Name: "test", Value: ptr("test1")},
+					},
+				},
+			},
+			grpcServerOpts: func() ([]grpc.ServerOption, error) {
+				return []grpc.ServerOption{}, nil
+			},
+		},
+		{
+			name: "with TLS config",
+			args: args{
+				ctx: context.Background(),
+				otlpConfig: &OTLP{
+					Protocol:    ptr("grpc"),
+					Compression: ptr("gzip"),
+					Timeout:     ptr(5000),
+					Certificate: ptr("testdata/server-certs/server.crt"),
+					Headers: []NameStringValuePair{
+						{Name: "test", Value: ptr("test1")},
+					},
+				},
+			},
+			grpcServerOpts: func() ([]grpc.ServerOption, error) {
+				opts := []grpc.ServerOption{}
+				tlsCreds, err := credentials.NewServerTLSFromFile("testdata/server-certs/server.crt", "testdata/server-certs/server.key")
+				if err != nil {
+					return nil, err
+				}
+				opts = append(opts, grpc.Creds(tlsCreds))
+				return opts, nil
+			},
+		},
+		{
+			name: "with TLS config and client key",
+			args: args{
+				ctx: context.Background(),
+				otlpConfig: &OTLP{
+					Protocol:          ptr("grpc"),
+					Compression:       ptr("gzip"),
+					Timeout:           ptr(5000),
+					Certificate:       ptr("testdata/server-certs/server.crt"),
+					ClientKey:         ptr("testdata/client-certs/client.key"),
+					ClientCertificate: ptr("testdata/client-certs/client.crt"),
+					Headers: []NameStringValuePair{
+						{Name: "test", Value: ptr("test1")},
+					},
+				},
+			},
+			grpcServerOpts: func() ([]grpc.ServerOption, error) {
+				opts := []grpc.ServerOption{}
+				cert, err := tls.LoadX509KeyPair("testdata/server-certs/server.crt", "testdata/server-certs/server.key")
+				if err != nil {
+					return nil, err
+				}
+				caCert, err := os.ReadFile("testdata/ca.crt")
+				if err != nil {
+					return nil, err
+				}
+				caCertPool := x509.NewCertPool()
+				caCertPool.AppendCertsFromPEM(caCert)
+				tlsCreds := credentials.NewTLS(&tls.Config{
+					Certificates: []tls.Certificate{cert},
+					ClientCAs:    caCertPool,
+					ClientAuth:   tls.RequireAndVerifyClientCert,
+				})
+				opts = append(opts, grpc.Creds(tlsCreds))
+				return opts, nil
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			n, err := net.Listen("tcp", "localhost:0")
+			require.NoError(t, err)
+
+			// this is a workaround, as providing 127.0.0.1 resulted in an "invalid URI for request" error
+			tt.args.otlpConfig.Endpoint = ptr(strings.ReplaceAll(n.Addr().String(), "127.0.0.1", "localhost"))
+
+			serverOpts, err := tt.grpcServerOpts()
+			require.NoError(t, err)
+
+			startGRPCLogsCollector(t, n, serverOpts)
+
+			exporter, err := otlpGRPCLogExporter(tt.args.ctx, tt.args.otlpConfig)
+			require.NoError(t, err)
+
+			logFactory := sdklogtest.RecordFactory{
+				Body: log.StringValue("test"),
+			}
+
+			assert.EventuallyWithT(t, func(collect *assert.CollectT) {
+				assert.NoError(collect, exporter.Export(context.Background(), []sdklog.Record{
+					logFactory.NewRecord(),
+				}))
+			}, 10*time.Second, 1*time.Second)
+		})
+	}
+}
+
+// grpcLogsCollector is an OTLP gRPC server that collects all requests it receives.
+type grpcLogsCollector struct {
+	collogpb.UnimplementedLogsServiceServer
+}
+
+var _ collogpb.LogsServiceServer = (*grpcLogsCollector)(nil)
+
+// startGRPCLogsCollector returns a *grpcLogsCollector that is listening at the provided
+// endpoint.
+//
+// If endpoint is an empty string, the returned collector will be listening on
+// the localhost interface at an OS chosen port.
+func startGRPCLogsCollector(t *testing.T, listener net.Listener, serverOptions []grpc.ServerOption) {
+	srv := grpc.NewServer(serverOptions...)
+	c := &grpcLogsCollector{}
+
+	collogpb.RegisterLogsServiceServer(srv, c)
+	go func() { _ = srv.Serve(listener) }()
+
+	t.Cleanup(func() {
+		srv.Stop()
+	})
+}
+
+// Export handles the export req.
+func (c *grpcLogsCollector) Export(
+	_ context.Context,
+	_ *collogpb.ExportLogsServiceRequest,
+) (*collogpb.ExportLogsServiceResponse, error) {
+	return &collogpb.ExportLogsServiceResponse{}, nil
 }